refactor: change stream controlling
