+++ backend/server.js
@@ -6,7 +6,7 @@ const ChatSocketHandler = require('./socket/ChatSocketHandler');
 const ChannelSocketHandler = require('./socket/ChannelSocketHandler');
 
 const channelController = require('./controllers/ChannelController');
-const streamController = require('./controllers/StreamController');
+const streamController = require('./services/streaming/StreamController');
 const ChannelService = require('./services/ChannelService');
 
 dotenv.config();
@@ -23,10 +23,10 @@ app.use('/api/channels', apiRouter);
 
 const PORT = 5000;
 const server = app.listen(PORT, () => {
-    console.log(`Server listening on Port ${PORT}`);
-    if(ChannelService.getCurrentChannel().restream) {
-        streamController.start();
-    }
+  console.log(`Server listening on Port ${PORT}`);
+  if (ChannelService.getCurrentChannel().restream) {
+    streamController.start(process.env.DEFAULT_CHANNEL_URL);
+  }
 });
